requirements.txt
django>=1.7.0
django-jsonfield==0.9.13
django-modeldict>=1.5.0
mock==1.0.1
